Payload-Encryptor/requirements.txt
typer
pycryptodomex
numpy
nuitka